+++ src/main/groovy/org/jetbrains/intellij/dependency/BuiltinPluginsRegistry.groovy
@@ -1,6 +1,7 @@
 package org.jetbrains.intellij.dependency
 
 import groovy.transform.ToString
+import groovy.xml.MarkupBuilder
 import org.jetbrains.annotations.NotNull
 import org.jetbrains.annotations.Nullable
 import org.jetbrains.intellij.Utils
@@ -17,17 +18,70 @@ class BuiltinPluginsRegistry implements Serializable {
     }
 
     static BuiltinPluginsRegistry fromDirectory(File pluginsDirectory, def loggingContext) {
-        //todo: save to file/load from file
         def result = new BuiltinPluginsRegistry(pluginsDirectory)
+        if (!result.fillFromCache(loggingContext)) {
+            Utils.debug(loggingContext, "Builtin registry cache is missing")
+            result.fillFromDirectory(loggingContext)
+            result.dumpToCache(loggingContext)
+        }
+        return result
+    }
+
+    private boolean fillFromCache(def loggingContext) {
+        def cache = cacheFile()
+        if (cache == null) return false
+
+        Utils.debug(loggingContext, "Builtin registry cache is found. Loading from $cache")
+        try {
+            Utils.parseXml(cache).children().forEach { node ->
+                plugins.put(node.id, new Plugin(node.id, node.directoryName, node.implementationDetail))
+            }
+            return true
+        }
+        catch (Throwable t) {
+            Utils.warn(loggingContext, "Cannot read builtin registry cache", t)
+            return false
+        }
+    }
+
+    private void fillFromDirectory(loggingContext) {
         def files = pluginsDirectory.listFiles()
         if (files != null) {
             for (file in files) {
                 if (file.isDirectory()) {
-                    result.add(file, loggingContext)
+                    add(file, loggingContext)
                 }
             }
         }
-        return result
+        Utils.debug(loggingContext, "Builtin registry populated with ${plugins.size()} plugins")
+    }
+
+    private void dumpToCache(def loggingContext) {
+        Utils.debug(loggingContext, "Dumping cache for builtin plugin")
+        def cacheFile = cacheFile()
+        def writer = null
+        try {
+            writer = new FileWriter(cacheFile)
+            new MarkupBuilder(writer).plugins {
+                for (p in plugins) {
+                    plugin (
+                        id: p.key,
+                        directoryName: p.value.directoryName,
+                        implementationDetail: p.value.implementationDetail,
+                    )
+                }
+            }
+        } catch (Throwable t) {
+            Utils.warn(loggingContext, "Failed to dump cache for builtin plugin", t)
+        } finally {
+            if (writer != null) {
+                writer.close()
+            }
+        }
+    }
+
+    private File cacheFile() {
+        new File(pluginsDirectory, "builtinRegistry.xml")
     }
 
     @Nullable
@@ -46,7 +100,7 @@ class BuiltinPluginsRegistry implements Serializable {
     }
 
     def add(@NotNull File artifact, def loggingContext) {
-        Utils.debug(loggingContext, "Adding  directory to plugins index: $artifact)")
+        Utils.debug(loggingContext, "Adding directory to plugins index: $artifact)")
         def intellijPlugin = Utils.createPlugin(artifact, false, loggingContext)
         if (intellijPlugin != null) {
             def plugin = new Plugin(intellijPlugin.pluginId, artifact.name, intellijPlugin.implementationDetail)